separate geo-location-list
