+++ catalystwan/endpoints/configuration/policy/list/geo_location.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import GeoLocationList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.geo_location import GeoLocationList, GeoLocationListEditPayload, GeoLocationListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class GeoLocationListEditPayload(GeoLocationList, PolicyListId):
-    pass
-
-
-class GeoLocationListInfo(GeoLocationList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyGeoLocationList(APIEndpoints, PolicyListEndpoints):
